[#40] Allow multiple commands to be run before saving (#42)
Allow multiple commands to be run

If an action from a gui part requires more than
one command to be executed, the current 
implementation requires a save after each command
execution. 

This change allows for lists of commands per action
to only be saved at the end of the list of commands.