[#57] Fixes wiping of old save files when template is changed (#63)
Fixes wiping of old save files when template is changed

Fixes the issue by allowing partial loading of a class if 
certain fields are missing. This will prevent future 
problems as it is likely the save template will be changed.

Additionally, as we are accepting broken test files, some 
additional safeguards have been added to missing 
key orderings and missing blocks.